Rename AddressesPaginator to Paginator
The addresses response has the key addresses, where the get-areas
has the results
