+++ lib/cyprus_postal_codes.rb
@@ -6,7 +6,7 @@ module CyprusPostalCodes
   autoload :Client, "cyprus_postal_codes/client"
   autoload :Error, "cyprus_postal_codes/error"
   autoload :Addresses, "cyprus_postal_codes/client/addresses"
-  autoload :AddressesPaginator, "cyprus_postal_codes/client/addresses_paginator"
+  autoload :Paginator, "cyprus_postal_codes/client/paginator"
   autoload :Areas, "cyprus_postal_codes/client/areas"
   autoload :Districts, "cyprus_postal_codes/client/districts"
 end
