+++ lib/cyprus_postal_codes/client.rb
@@ -6,7 +6,7 @@ require "faraday_middleware"
 module CyprusPostalCodes
   class Client
     include Addresses
-    include AddressesPaginator
+    include Paginator
     include Areas
     include Districts
