+++ lib/cyprus_postal_codes/client/areas.rb
@@ -3,8 +3,8 @@
 module CyprusPostalCodes
   class Client
     module Areas
-      def areas(district:)
-        get("get-areas", district: district)
+      def areas(district:, page_token: nil)
+        get("get-areas", district: district, page_token: page_token)
       end
     end
   end
